requirements.txt
# SportsAnalytics-CV Requirements
# Author: Malav Patel
# Last Updated: February 2026

# Core ML/DL
ultralytics>=8.0.0
torch>=2.1.0
torchvision>=0.16.0

# Computer Vision (headless only — no GUI deps in CI/Docker)
opencv-python-headless>=4.8.0
supervision>=0.18.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Demo UI
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
pillow>=10.1.0

# Roboflow (for dataset management)
roboflow>=1.1.0
